Add development protobuf definition for process context
This PR adds the protobuf definition for the specification proposed in
<https://docs.google.com/document/d/1-4jo29vWBZZ0nKKAOG13uAQjRcARwmRc4P313LTbPOE/edit?tab=t.0>.

The `OtelProcessCtx` message is used to allow in-process libraries
(such as the OTel SDKs running inside an application) to share
information with outside readers (such as the OpenTelemetry eBPF
Profiler).

At this point, it's not clear to me if this is the final home for
this file/the specification, but it looks like a good home for now
while we develop and iterate on it.