Fix support for modern kernels without PR_SET_VMA_ANON_NAME
As uncovered during review, it's possible that even on modern kernels,
we're not able to name the process context mapping.

This is not a big problem since we already had a fallback for older
kernels, so I've tweaked the code so that the fallback is also usable
on modern kernels (e.g. no more peeking at kernel version when deciding
to use the fallback or not).

This doesn't impact publishing (it already accounted for naming possibly
failing), but I've updated the testing reader inside
`otel_process_ctx.c` as well as the one in `otel_process_ctx_dump.sh`.